feat: refactor and consolidate milestone dialog hooks into useChangeMilestoneStatusDialogHook
