+++ src/components/modules/escrow/ui/dialogs/hooks/change-status-escrow-dialog.hook.ts
@@ -2,16 +2,27 @@
 
 "use client";
 
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { z } from "zod";
 import {
   useGlobalAuthenticationStore,
   useGlobalBoundedStore,
 } from "@/core/store/data";
 import { useEscrowUIBoundedStore } from "../../../store/ui";
-import { Escrow, Milestone } from "@/@types/escrow.entity";
-import { changeMilestoneStatus } from "../../../services/change-milestone-status.service";
 import { toast } from "@/hooks/toast.hook";
+import { formSchema } from "../../../schema/complete-milestone.schema";
+import { useEscrowBoundedStore } from "../../../store/data";
+import { changeMilestoneStatus } from "../../../services/change-milestone-status.service";
+import { ChangeMilestoneStatusPayload } from "@/@types/escrow.entity";
 
-const useChangeStatusEscrowDialog = () => {
+interface changeMilestoneStatusDialogHook {
+  setIsCompleteMilestoneDialogOpen: (value: boolean) => void;
+}
+
+const useChangeMilestoneStatusDialogHook = ({
+  setIsCompleteMilestoneDialogOpen,
+}: changeMilestoneStatusDialogHook) => {
   const { address } = useGlobalAuthenticationStore();
   const setIsChangingStatus = useEscrowUIBoundedStore(
     (state) => state.setIsChangingStatus,
@@ -22,35 +33,54 @@ const useChangeStatusEscrowDialog = () => {
   const setSelectedEscrow = useGlobalBoundedStore(
     (state) => state.setSelectedEscrow,
   );
+  const selectedEscrow = useGlobalBoundedStore((state) => state.selectedEscrow);
+  const completingMilestone = useEscrowBoundedStore(
+    (state) => state.completingMilestone,
+  );
+  const milestoneIndex = useEscrowBoundedStore((state) => state.milestoneIndex);
   const fetchAllEscrows = useGlobalBoundedStore(
     (state) => state.fetchAllEscrows,
   );
   const activeTab = useEscrowUIBoundedStore((state) => state.activeTab);
 
-  const changeMilestoneStatusSubmit = async (
-    selectedEscrow: Escrow,
-    milestone: Milestone,
-    index: number,
-  ) => {
-    setIsChangingStatus(true);
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      evidence: "",
+    },
+    mode: "onChange",
+  });
 
+  const onSubmit = async ({
+    evidence,
+  }: Pick<ChangeMilestoneStatusPayload, "evidence">) => {
+    setIsChangingStatus(true);
+    console.log({
+      contractId: selectedEscrow?.contractId,
+      milestoneIndex: milestoneIndex?.toString() || "0",
+      newStatus: "completed",
+      serviceProvider: address,
+      evidence,
+    });
     try {
       const response = await changeMilestoneStatus({
         contractId: selectedEscrow?.contractId,
-        milestoneIndex: index.toString(),
+        milestoneIndex: milestoneIndex?.toString() || "0",
         newStatus: "completed",
         serviceProvider: address,
+        evidence,
       });
 
       if (response.status === "SUCCESS") {
         setIsChangingStatus(false);
         setIsDialogOpen(false);
+        setIsCompleteMilestoneDialogOpen(false);
         setSelectedEscrow(undefined);
         fetchAllEscrows({ address, type: activeTab || "serviceProvider" });
 
         toast({
           title: "Success",
-          description: `The Milestone ${milestone.description} has been completed.`,
+          description: `The Milestone ${completingMilestone?.description} has been completed.`,
         });
       }
     } catch (error: any) {
@@ -64,7 +94,16 @@ const useChangeStatusEscrowDialog = () => {
     }
   };
 
-  return { changeMilestoneStatusSubmit };
+  const handleClose = () => {
+    setIsCompleteMilestoneDialogOpen?.(false);
+  };
+
+  return {
+    form,
+    onSubmit,
+    handleClose,
+    setIsDialogOpen,
+  };
 };
 
-export default useChangeStatusEscrowDialog;
+export default useChangeMilestoneStatusDialogHook;
